Python: Averages.read: warning when both 1D and 2D averages are read.
If the times of the 1D and 2D averages differ, the returned averages
object may get assigned any one of them as the 't' attribute (with the
times of the other kind of averages not being returned at all). I have
changed the code to save the time as an attribute to each plane.