diagnostics.f90: ysum_mn_name_xz_npar: fix normalization of the average for non-Cartesian coordinates
The fact that the line element for the y coordinate in
cylindrical/spherical coordinates is x*dy was accounted for while
summing the values, but not while normalizing the final result to get
the average. I have now ensured that both are handled in a consistent
way to give the correct result.