Add sample: conv-slab_cp_2
A variant of conv-slab that sets cp=2. The idea is to test for implicit
assumptions that cp==1 in various parts of the code that are used in
convection setups.

With my recent fixes, the results of this are identical to those of the
conv-slab sample (as long as you consider ss/cp rather than ss).