eos_idealgas_vapor: calculate cp etc in eos_before_boundary subroutine
Subroutine get_gamma_etc requires cp to be in the f-array. Usage of
get_gamma_etc in boundary condition routines then requires that cp be
present in the f-array before the boundary conditions are called.
Calculating cp requires fvap and mumol1; we then keep them in the
f-array
to avoid unnecessary recomputation if these are needed as pencils later
on.