eos_idealgas/bc_ss_temp_z: remove l2nd=.true. from the set_ghosts_for_onesided_ders calls
Verified that fradz_Kprof is now constant in the steady state of a 1D
thermal conduction problem.

Matthias, do you remember why you initially set this to .true.? From
set_ghosts_for_onesided_ders, it looks like the only effect of setting
it to .true. is to leave one of the ghost points unset, but am I missing
something?