[cherry-pick] Python: test_read_var: reduce precision
It makes no sense to test output from a single-precision run to a
precision of 1e-9.

Backstory:

I just noticed that the np.std(f[1]) test fails depending on compiler
options. The reference value corresponds to -O3 (on gfortran), while the
value is different with -O1. After many painful hours of debugging
(during which I considered burning all my electronic devices and
retreating to the mountains), I found the responsible options were
-fvect-cost-model=dynamic -ftree-loop-vectorize.

This test reads data from a single-precision sample, which probably
explains why one gets an absolute change of the order of 2e-9 depending
on whether vectorization is enabled.

I have checked that if REAL_PRECISION is set to double, np.std(f[1]) is
unchanged (to within an absolute precision of 1e-9) by vectorization.