pc_auto-test: support for using timeout rather than bin/reaper
This only affects cases where a time limit is specified. reaper used an
approach of just killing all processes with a specified name after the
time limit, which can cause problems if multiple pc_auto-test instances
are running (even with different $PENCIL_HOME). If the timeout utility
is available, we use that instead, which I think would avoid this
limitation.

timeout is part of GNU coreutils. Can someone with a Mac check if it is
available there as well?