pencil-test: also resolve log_dir and previous_dir to absolute paths
If they are left as relative paths, the log directory is created
relative to $PENCIL_HOME (see build_remote_cmd). This does not sound
like something anyone would want.

The new behaviour is that if a relative path is specified, it is with
respoct to the directory in which pencil-test was invoked.

I had made a similar change in pc_auto-test earlier, but hadn't done it
here. It seems strange to have the code handling this option duplicated
in both pc_auto-test and pencil-test. At some point, should try to make
these use a common code path.