diagnostics: xyaverages_z: handle the ilabel_save case
This is the default value of itype_name_z (see allocate_xyaverages). Not
handling it leads to an error when the dcoolz diagnostic is present in
xyaver.in and lcooling_general=T. Expected behaviour is that the value
of the diagnostic is simply zero in this case, since it is not
implemented in entropy.f90/get_cool_general.