Python: pc.read.pvar: split dat reading logic into separate function
Ugh. Why repeat the same block of code in three places?

This is currently broken (seems to read an array of the wrong size in
the helical-MHDTurb sample); this refactor is so that I can get the code
into a form that is easy to understand and debug.

Fred, you coded this function. Did you test the io_dist codepath? If so,
please let me know what kind of setup it works for (if at all).