Add sample: HDF5 version of helical-MHDTurb
Mainly so that we can also test writing and reading (in Python) the HDF5
averages.

For unknown reasons, the 2D averages (yaver and zaver) do not match
those in the helical-MHDTurb sample. The current reference data are the
same as in the standard (nohdf5, nompicomm) sample, and thus this test
currently fails (running script tests).